+++ fighthealthinsurance/views.py
@@ -263,62 +263,6 @@ class ChooseAppeal(View):
             print(form)
 
 
-class FaxFollowUpView(generic.FormView):
-    template_name = "faxfollowup.html"
-    form_class = FaxResendForm
-
-    def get_initial(self):
-        # Set the initial arguments to the form based on the URL route params.
-        return self.kwargs
-
-    def form_valid(self, form):
-        FollowUpHelper.store_follow_up_result(**form.cleaned_data)
-        return render(self.request, "fax_followup_thankyou.html")
-
-
-class SendFaxView(View):
-
-    def get(self, request, **kwargs):
-        SendFaxHelper.remote_send_fax(**self.kwargs)
-        return render(self.request, "fax_thankyou.html")
-
-
-class StageFaxView(generic.FormView):
-    form_class = FaxForm
-
-    def form_valid(self, form):
-        staged = SendFaxHelper.stage_appeal_fax(**form.cleaned_data)
-        stripe.api_key = settings.STRIPE_API_SECRET_KEY
-        stripe.publishable_key = settings.STRIPE_API_PUBLISHABLE_KEY
-        product = stripe.Product.create(name="Fax")
-        product_price = stripe.Price.create(
-            unit_amount=500, currency="usd", product=product["id"]
-        )
-        items = [
-            {
-                "price": product_price["id"],
-                "quantity": 1,
-            }
-        ]
-        checkout = stripe.checkout.Session.create(
-            line_items=items,
-            mode="payment",  # No subscriptions
-            success_url=self.request.build_absolute_uri(
-                reverse(
-                    "sendfaxview",
-                    kwargs={
-                        "uuid": staged.uuid,
-                        "hashed_email": staged.hashed_email,
-                    },
-                ),
-            ),
-            cancel_url=self.request.build_absolute_uri(reverse("root")),
-            customer_email=form.cleaned_data["email"],
-        )
-        checkout_url = checkout.url
-        return redirect(checkout_url)
-
-
 class GenerateAppeal(View):
     def post(self, request):
         form = DenialRefForm(request.POST)
